Add files via upload
Here I created a snapshot for volume of an instance and I connected it to different instance.

=================================================================

To Create a snapshot, Go to the volume select the specific drive in the volume -> Actions -> Create snapshot  

Give description -> create snapshot

Now the snapshot is created
--------------------------------------------------------------------------------------------------

We're going to use the snapshot to back up the drive 

Create a file small file in the drive you created snapshot and delete the instance and the volumes as well because only root volume will be delete if we delete the instance. We have to manually delete the added volume 

Create a new instance in configuration storage tab -> Adavance

Click on snapshot -> select your snapshot -> size has to match snapshot drive size(drive has 12 gb you have to increase 4gb here)

In practice : Also keep delete on termination YES  otherwise you have to contact customer service to delete it 

After that click on launch instance, now you can see the back up drive in the running instance.